fixed bugs

+++ song-microservice/src/main/java/com/csc301/songmicroservice/SongController.java
@@ -90,9 +90,9 @@ public class SongController {
 		
 		DbQueryStatus status = songDal.deleteSongById(songId);
 		
-	//	if(status.getdbQueryExecResult() == DbQueryExecResult.QUERY_OK) {
+		if(status.getdbQueryExecResult() == DbQueryExecResult.QUERY_OK) {
 			restTemplate.exchange("http://localhost:3002/deleteAllSongsFromDb/" + songId, HttpMethod.PUT, null, String.class);
-	//	} 
+		} 
 		
 
 		
@@ -115,7 +115,9 @@ public class SongController {
 		if(params.get(Song.KEY_SONG_NAME) != null && params.get(Song.KEY_SONG_ARTIST_FULL_NAME) != null && params.get(Song.KEY_SONG_ALBUM) != null) {
 			Song _song = new Song(params.get(Song.KEY_SONG_NAME), params.get(Song.KEY_SONG_ARTIST_FULL_NAME), params.get(Song.KEY_SONG_ALBUM));
 			status = songDal.addSong(_song);
-			restTemplate.exchange("http://localhost:3002/addSongToDB/" + _song.getId(), HttpMethod.PUT, null, String.class);
+			if(status.getdbQueryExecResult() == DbQueryExecResult.QUERY_OK) {
+				restTemplate.exchange("http://localhost:3002/addSongToDB/" + _song.getId(), HttpMethod.PUT, null, String.class);
+			} 
 		}
 		else {
 			status = new DbQueryStatus("missing paramater", DbQueryExecResult.QUERY_ERROR_NOT_FOUND);
